fix: address code review feedback for effort level refactor
Fix gateway effort passthrough, remove dead thinking-detector code,
rename ThinkingModeSelector to EffortSelector, fix stale effort state
in session panel, and update spawn template references.
