+++ apps/agor-ui/src/components/EffortSelector/EffortSelector.tsx
