--- apps/agor-ui/src/components/ThinkingModeSelector/index.ts
@@ -1,2 +1,0 @@
-export type { EffortLevel } from '@agor-live/client';
-export { EffortSelector } from './ThinkingModeSelector';
